lab-express-rest-show-create - DONE
